move clipboard copy to beta folder, create deprecated version
