+++ app/components/primer/beta/clipboard_copy.html.erb
